feat: Change register tree screen
Changed register tree screen to be a floating card.

Moved components from TreeMapScreen file to their respective files.

Deleted RegisterTreeScreen file.

Removed some lost console.log().
