+++ src/navigation/AppNavigator.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 
-import RegisterTreeScreen from "../screens/RegisterTreeScreen";
 import TreeMapScreen from "../screens/TreeMapScreen";
 import routes from "./routes";
 
@@ -15,7 +14,6 @@ const AppNavigator = () => {
         component={TreeMapScreen}
         options={{ headerShown: false }}
       />
-      <Stack.Screen name={routes.REGISTER_TREE} component={RegisterTreeScreen} />
     </Stack.Navigator>
   );
 };
