+++ src/navigation/routes.js
@@ -1,7 +1,5 @@
 export default Object.freeze({
   WELCOME: "Welcome",
   REGISTER_ACCOUNT: "Register Account",
-  REGISTER_TREE: "Register Tree",
   MAP: "Map",
-  NAVIGATOR: "Navigator"
 });
